[cudamapper] Perf improvements for cudamapper

+++ cudamapper/include/claragenomics/cudamapper/types.hpp
@@ -35,7 +35,7 @@ using position_in_read_t = std::uint32_t;
 /// representation_t
 using representation_t = std::uint64_t; // this depends on kmer size, in some cases could also be 32-bit
 /// read_id_t
-using read_id_t = std::uint64_t; // can this be 32-bit?
+using read_id_t = std::uint32_t;
 
 /// Relative strand - represents whether query and target
 /// are on the same DNA strand (i.e Forward) or not (i.e Reverse).
